requirements.txt
fastapi\openai
umpy
